The Matrix-t distribution.

Density and sampling for the Matrix-t distribution.

dMT(X, Lambda, SigmaR, SigmaC, nu, log = FALSE)

rMT(n, Lambda, SigmaR, SigmaC, nu)

Arguments

X

Argument to the density function. Either a p x q matrix or a p x q x n array.

Lambda

Mean parameter. Either a p x q matrix or a p x q x n array.

SigmaR

Between-row covariance matrix. Either a p x p matrix or a p x p x n array.

SigmaC

Between-column covariance matrix Either a q x q matrix or a q x q x n array.

nu

Degrees-of-freedom parameter. A scalar or vector of length n.

log

Logical; whether or not to compute the log-density.

n

Integer number of random samples to generate.

Value

A vector length n for density evaluation, or an array of size p x q x n for random sampling.

Details

The Matrix-T distribution \(\boldsymbol{X} \sim \textrm{Matrix-T}(\boldsymbol{\Lambda}, \boldsymbol{\Sigma}, \boldsymbol{\Psi}, \nu)\) on a random matrix \(\boldsymbol{X}_{p \times q}\) is the marginal distribution of \(\boldsymbol{X}\) in \((\boldsymbol{X}, \boldsymbol{V}) \sim \textrm{MNIW}(\boldsymbol{\Lambda}, \boldsymbol{\Sigma}, \boldsymbol{\Psi}, \nu)\), where the Matrix-Normal Inverse-Wishart (MNIW) distribution is defined in MNIW-dist.